Fix memory leak, consolidate ballot batch and paper into 1 column

+++ src/main/java/au/id/tmm/hypotheticalsenate/database/DataSource.java
@@ -1,20 +1,15 @@
 package au.id.tmm.hypotheticalsenate.database;
 
 import au.com.bytecode.opencsv.CSVReader;
-import au.id.tmm.hypotheticalsenate.model.AustralianState;
-import com.google.common.collect.ImmutableMap;
 import org.apache.commons.io.FileUtils;
 import org.apache.commons.io.FilenameUtils;
 
-import javax.annotation.Nullable;
 import java.io.BufferedInputStream;
 import java.io.File;
 import java.io.FileInputStream;
 import java.io.IOException;
 import java.io.InputStream;
 import java.io.InputStreamReader;
-import java.net.MalformedURLException;
-import java.net.URL;
 import java.util.Enumeration;
 import java.util.Optional;
 import java.util.zip.ZipEntry;
@@ -23,58 +18,17 @@ import java.util.zip.ZipFile;
 /**
  * @author timothy
  */
-public enum DataSource {
-
-    SENATE_CANDIDATES("http://results.aec.gov.au/17496/Website/Downloads/SenateCandidatesDownload-17496.txt"),
-    GROUP_VOTING_TICKETS("http://results.aec.gov.au/17496/Website/Downloads/SenateGroupVotingTicketsDownload-17496.txt"),
-    GROUP_FIRST_PREFERENCES("http://results.aec.gov.au/17496/Website/Downloads/SenateFirstPrefsByStateByGroupByVoteTypeDownload-17496.txt"),
-//    ACT_BTL_PREFERENCES(),
-//    NSW_BTL_PREFERENCES(),
-    NT_BTL_PREFERENCES("http://results.aec.gov.au/17496/Website/External/SenateStateBtlDownload-17496-NT.zip", true, "SenateStateBTLPreferences-17496-NT.txt"),
-//    QLD_BTL_PREFERENCES(),
-//    SA_BTL_PREFERENCES(),
-//    TAS_BTL_PREFERENCES(),
-    VIC_BTL_PREFERENCES("http://results.aec.gov.au/17496/Website/External/SenateStateBtlDownload-17496-VIC.zip", true, "SenateStateBTLPreferences-17496-VIC.txt"),
-//    WA_BTL_PREFERENCES()
-    ;
-
-    public static final ImmutableMap<AustralianState, DataSource> BTL_DATA_MAP =
-            ImmutableMap.<AustralianState, DataSource>builder()
-                    .put(AustralianState.VIC, VIC_BTL_PREFERENCES)
-                    .put(AustralianState.NT, NT_BTL_PREFERENCES)
-                    .build();
-
-    public static final String DOWNLOADS_LOCATION = "AEC_DATA/";
+public class DataSource {
 
     public static final int CONNECTION_TIMEOUT = 10000;
     public static final int READ_TIMEOUT = 10000;
 
-    private URL downloadLocation;
-    private boolean inZip;
-    private Optional<String> zipEntryName = Optional.empty();
-    private char csvSeparator;
+    private AECResource aecResource;
+    private File downloadDirectory;
 
-    private DataSource(String downloadLocation, boolean inZip, @Nullable String zipEntryName, char csvSeparator) {
-        try {
-            this.downloadLocation = new URL(downloadLocation);
-        } catch (MalformedURLException e) {
-            throw new RuntimeException(e);
-        }
-        this.inZip = inZip;
-        this.zipEntryName = Optional.ofNullable(zipEntryName);
-        this.csvSeparator = csvSeparator;
-    }
-
-    private DataSource(String downloadLocation, boolean inZip, @Nullable String zipEntryName) {
-        this(downloadLocation, inZip, zipEntryName, '\t');
-    }
-
-    private DataSource(String downloadLocation, char csvSeparator) {
-        this(downloadLocation, false, null, csvSeparator);
-    }
-
-    private DataSource(String downloadLocation) {
-        this(downloadLocation, '\t');
+    public DataSource(AECResource aecResource, File downloadDirectory) {
+        this.aecResource = aecResource;
+        this.downloadDirectory = downloadDirectory;
     }
 
     public boolean isDownloaded() {
@@ -83,24 +37,30 @@ public enum DataSource {
     }
 
     private File getExpectedLocalFile() {
-        return new File(DOWNLOADS_LOCATION, FilenameUtils.getName(downloadLocation.getPath()));
+        return new File(downloadDirectory, FilenameUtils.getName(this.aecResource.getResourceLocation().getPath()));
     }
 
     public void download() {
         try {
+            if (!downloadDirectory.isDirectory()) {
+                throw new RuntimeException("The download directory is not a directory");
+            }
+
             File localFile = this.getExpectedLocalFile();
 
             localFile.delete();
 
-            FileUtils.copyURLToFile(this.downloadLocation, localFile, CONNECTION_TIMEOUT, READ_TIMEOUT);
+            FileUtils.copyURLToFile(this.aecResource.getResourceLocation(), localFile,
+                    CONNECTION_TIMEOUT, READ_TIMEOUT);
         } catch (IOException e) {
-            throw new RuntimeException("An error occurred while downloading " + this.downloadLocation.toString(), e);
+            throw new RuntimeException("An error occurred while downloading "
+                    + this.aecResource.getResourceLocation().toString(), e);
         }
     }
 
     public InputStream openInputStream() {
         try {
-            if (this.inZip) {
+            if (this.aecResource.isInZip()) {
                 return inputStreamFromZip();
             } else {
                 return new FileInputStream(this.getExpectedLocalFile());
@@ -112,9 +72,9 @@ public enum DataSource {
     }
 
     private InputStream inputStreamFromZip() throws IOException {
-        if (!this.zipEntryName.isPresent()) {
+        if (!this.aecResource.getZipEntryName().isPresent()) {
             throw new RuntimeException("A data source expected to be found inside a zip file does not " +
-                            "specify the expected name within the archive");
+                    "specify the expected name within the archive");
         }
 
         ZipFile zipFile = new ZipFile(this.getExpectedLocalFile());
@@ -124,7 +84,7 @@ public enum DataSource {
 
         while(zipEntries.hasMoreElements()) {
             ZipEntry currentEntry = zipEntries.nextElement();
-            if (currentEntry.getName().equals(this.zipEntryName.get())) {
+            if (currentEntry.getName().equals(this.aecResource.getZipEntryName().get())) {
                 matchingEntry = Optional.of(currentEntry);
                 break;
             }
@@ -133,12 +93,13 @@ public enum DataSource {
         if (matchingEntry.isPresent()) {
             return zipFile.getInputStream(matchingEntry.get());
         } else {
-            throw new RuntimeException("The expected entry " + this.zipEntryName.get() + " was not found in " +
-                    "the archive " + zipFile.getName());
+            throw new RuntimeException("The expected entry " + this.aecResource.getZipEntryName().get()
+                    + " was not found in " + "the archive " + zipFile.getName());
         }
     }
 
     public CSVReader getCSVReader() {
-        return new CSVReader(new InputStreamReader(new BufferedInputStream(this.openInputStream())), this.csvSeparator);
+        return new CSVReader(new InputStreamReader(new BufferedInputStream(this.openInputStream())),
+                this.aecResource.getCsvSeparator());
     }
 }
